Moved CountrySelectorUtils to searchCountryOptions
